requirements.txt
# Python version requirement: >= 3.9, < 3.12 (Modified for Python 3.13 compatibility)

# Core dependencies
python-dotenv>=0.19.0,<1.0.0

# Deep Learning - Using compatible versions
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.21.0,<5.0.0
accelerate>=0.12.0,<1.0.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0,<5.0.0

# Audio Processing
soundfile>=0.10.3,<1.0.0
# pydub>=0.25.1,<1.0.0 # Removed - Replaced with moviepy for audio manipulation
ffmpeg-python>=0.2.0,<1.0.0
silero-vad>=5.1.0,<6.0.0

# Video Processing
moviepy>=1.0.3,<2.0.0

# Voice Activity Detection - Commented out due to compatibility issues
# pyannote.audio>=2.1.0,<3.0.0

# Text-to-Speech - Using alternative TTS engines compatible with Python 3.13
# Option 3: Alternative TTS engines
gTTS>=2.3.0,<3.0.0  # Google Text-to-Speech
pyttsx3>=2.90,<3.0.0  # Offline TTS

# Progress bars and logging
tqdm>=4.62.0,<5.0.0

# Additional requirements
setuptools>=65.5.1
wheel>=0.38.0

# Development Tools / Static Analysis
mypy>=1.0.0,<2.0.0